better docstrings, add 403 error to inference.py
